feat: add main entry point
