+++ README.md
@@ -24,3 +24,17 @@ cp env.example .env
 - `TESSERACT_LANG` — язык для OCR Tesseract (например, `eng`);
 - `OUTPUT_DIR` — папка, куда будут сохраняться обработанные документы (например, `Archive`).
 
+## Запуск
+
+Основной скрипт расположен в `main.py`:
+
+```bash
+python main.py ./Input --output ./Archive --dry-run
+```
+
+Параметры:
+
+- `input_dir` — директория с исходными файлами;
+- `--output` — корневая папка для отсортированных документов (по умолчанию `Archive`);
+- `--dry-run` — показывать предполагаемые действия без перемещения файлов.
+
